+++ crates/nostr-sdk/src/client/mod.rs
@@ -1406,14 +1406,13 @@ impl Client {
     /// <https://github.com/nostr-protocol/nips/blob/master/59.md>
     #[inline]
     #[cfg(feature = "nip59")]
-    pub async fn gift_wrap<I, R>(
+    pub async fn gift_wrap<I>(
         &self,
         receiver: &PublicKey,
-        rumor: R,
+        rumor: UnsignedEvent,
         extra_tags: I,
     ) -> Result<Output<EventId>, Error>
     where
-        R: Into<EventBuilderOrUnsignedEvent>,
         I: IntoIterator<Item = Tag>,
     {
         // Acquire signer
@@ -1432,17 +1431,16 @@ impl Client {
     /// <https://github.com/nostr-protocol/nips/blob/master/59.md>
     #[inline]
     #[cfg(feature = "nip59")]
-    pub async fn gift_wrap_to<I, U, R, IT>(
+    pub async fn gift_wrap_to<I, U, IT>(
         &self,
         urls: I,
         receiver: &PublicKey,
-        rumor: R,
+        rumor: UnsignedEvent,
         extra_tags: IT,
     ) -> Result<Output<EventId>, Error>
     where
         I: IntoIterator<Item = U>,
         U: TryIntoUrl,
-        R: Into<EventBuilderOrUnsignedEvent>,
         IT: IntoIterator<Item = Tag>,
         pool::Error: From<<U as TryIntoUrl>::Err>,
     {